refactor: rename types module and update imports for validation system
This commit renames the `types.py` module to `validation_types.py` to better reflect its purpose within the ATS PDF Generator. Additionally, all relevant imports across the codebase have been updated to reference the new module name, ensuring consistency and clarity in the validation system's architecture.
