+++ src/ats_pdf_generator/validator/__init__.py
@@ -6,7 +6,7 @@ Each validator implements specific validation rules and returns violations.
 """
 
 # First-party
-from ats_pdf_generator.types import Violation
+from ats_pdf_generator.validation_types import Violation
 from ats_pdf_generator.validator.contact_validator import ContactValidator
 from ats_pdf_generator.validator.validator import EMOJI_PATTERN, validate_document
